+++ Warkey.Core/Models/SettingsModel.cs
