+++ Warkey.Infrastructure/Keyboard/Detector.cs
@@ -1,8 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Input;
 
 namespace Warkey.Infrastructure.Keyboard
 {
@@ -30,12 +33,15 @@ namespace Warkey.Infrastructure.Keyboard
         private const int WM_KEYUP = 0x0101;
         private const int WM_SYSKEYDOWN = 0x0104;
 
-        private LowLevelKeyboardProc _proc = HookCallback;
-        private IntPtr _hookID = IntPtr.Zero;
+        private static LowLevelKeyboardProc _proc = HookCallback;
+        private static IntPtr _hookID = IntPtr.Zero;
 
-        public void Initialize()
+        public static Func<bool> Precondition { get; set; }
+
+        public Detector(Func<bool> precondition)
         {
             _hookID = SetHook(_proc);
+            Precondition = precondition;
         }
 
         public void Dispose()
@@ -45,8 +51,8 @@ namespace Warkey.Infrastructure.Keyboard
 
         private IntPtr SetHook(LowLevelKeyboardProc proc)
         {
-            using (Process curProcess = Process.GetCurrentProcess())
-            using (ProcessModule curModule = curProcess.MainModule)
+            using (var curProcess = Process.GetCurrentProcess())
+            using (var curModule = curProcess.MainModule)
             {
                 return SetWindowsHookEx(WH_KEYBOARD_LL, proc,
                     GetModuleHandle(curModule.ModuleName), 0);
@@ -55,20 +61,21 @@ namespace Warkey.Infrastructure.Keyboard
 
         private delegate IntPtr LowLevelKeyboardProc(int nCode, IntPtr wParam, IntPtr lParam);
 
-        private IntPtr HookCallback(Action<bool> preprocessCheck, int nCode, IntPtr wParam, IntPtr lParam)
+        //preprocess: check iswar3forground and ischatting
+        private static IntPtr HookCallback(int nCode, IntPtr wParam, IntPtr lParam)
         {
             // iswar3foreground have to put here
             // because if not in foreground no need to do anything
             // else need to have pointer to mainservice iswar3foreground
             // and pass in here when initialize
             // which is not possible with C#
-            if (preprocessCheck && !ChatboxService.IsChatting && nCode >= 0)
+            if (Precondition() && nCode >= 0)
             {
                 int vkCode = Marshal.ReadInt32(lParam);
 
                 if (wParam == (IntPtr)WM_KEYDOWN)
                 {
-                    GlobalKeyDown(null, new HotkeyItem()
+                    GlobalKeyDown(null, new HotkeyModel()
                     {
                         Alt = false,
                         Key = KeyInterop.KeyFromVirtualKey(vkCode)
@@ -78,7 +85,7 @@ namespace Warkey.Infrastructure.Keyboard
                 {
                     if (vkCode != VK_LMENU)
                     {
-                        GlobalKeyDown(null, new HotkeyItem()
+                        GlobalKeyDown(null, new HotkeyModel()
                         {
                             Alt = true,
                             Key = KeyInterop.KeyFromVirtualKey(vkCode)
@@ -90,6 +97,13 @@ namespace Warkey.Infrastructure.Keyboard
             return CallNextHookEx(_hookID, nCode, wParam, lParam);
         }
 
-        public event EventHandler<HotkeyItem> GlobalKeyDown;
+        public static event EventHandler<HotkeyModel> GlobalKeyDown;
+
+        public class HotkeyModel
+        {
+            public bool Alt { get; set; }
+            public Key Key { get; set; }
+
+        }
     }
 }
