--- Warkey.Shared/ScreenResolution.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Warkey.Shared
-{
-    public class ScreenResolution
-    {
-        public int Width { get; set; }
-        public int Height { get; set; }
-    }
-}
